Miscellaneous code cleanup
1. Remove unused resources (Android Lint)
2. Resolve various code inspection checks in Android Studio
