+++ amplify-core/src/androidTest/java/com/amplifyframework/hub/HubInstrumentedTest.java
@@ -62,9 +62,8 @@ public final class HubInstrumentedTest {
     @Test
     public void subscriptionTokenCanBeUsedToUnsubscribe() throws InterruptedException {
         final CountDownLatch waitUntilSubscriptionIsReceived = new CountDownLatch(1);
-        final SubscriptionToken token = Amplify.Hub.subscribe(HubChannel.STORAGE, payload -> {
-            waitUntilSubscriptionIsReceived.countDown();
-        });
+        final SubscriptionToken token = Amplify.Hub.subscribe(HubChannel.STORAGE,
+            payload -> waitUntilSubscriptionIsReceived.countDown());
         assertNotNull(token);
         assertNotNull(token.getUuid());
 
@@ -138,7 +137,7 @@ public final class HubInstrumentedTest {
     public void multiplePublications() throws InterruptedException {
         final int numPublications = 10;
         final CountDownLatch allSubscriptionsReceived = new CountDownLatch(numPublications);
-        final List<Integer> subscriptionsReceived = new ArrayList<Integer>();
+        final List<Integer> subscriptionsReceived = new ArrayList<>();
 
         final SubscriptionToken token = Amplify.Hub.subscribe(HubChannel.STORAGE, payload -> {
             if (payload.getEventData() instanceof Integer) {
@@ -178,8 +177,8 @@ public final class HubInstrumentedTest {
         final int numPublications = 10;
         final int numDataTypes = 2;
         final CountDownLatch allSubscriptionsReceived = new CountDownLatch(numPublications);
-        final List<Integer> integerSubscriptionsReceived = new ArrayList<Integer>();
-        final List<String> stringSubscriptionsReceived = new ArrayList<String>();
+        final List<Integer> integerSubscriptionsReceived = new ArrayList<>();
+        final List<String> stringSubscriptionsReceived = new ArrayList<>();
         final String stringSubscriptionValue = "weatherAlwaysRemainsTheSame";
 
         final SubscriptionToken token = Amplify.Hub.subscribe(HubChannel.STORAGE, payload -> {
