+++ amplify-core/src/main/AndroidManifest.xml
@@ -14,5 +14,4 @@
    permissions and limitations under the License.
 -->
 
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.amazonaws.amplify.core" />
+<manifest package="com.amazonaws.amplify.core" />
