+++ amplify-core/src/main/java/com/amplifyframework/core/plugin/Plugin.java
@@ -45,5 +45,6 @@ public interface Plugin<E> extends CategoryTypeable {
      * Returns escape hatch for plugin to enable lower-level client use-cases.
      * @return the client used by category plugin
      */
+    @SuppressWarnings("SameReturnValue")
     E getEscapeHatch();
 }
