+++ amplify-core/src/main/java/com/amplifyframework/hub/HubPayloadFilter.java
@@ -20,8 +20,8 @@ import androidx.annotation.NonNull;
 /**
  * A filter which can be used to control which {@link HubPayload}s a
  * subscribed {@link HubListener} will receive from the Hub. The
- * {@link HubPaylods} utility constants factories for some commonly
- * used types of {@HubPayloadFilter}. Alternately, you are free to
+ * {@link HubFilters} utility constants factories for some commonly
+ * used types of {@link HubPayloadFilter}. Alternately, you are free to
  * implement your own custom, more complex, logic.
  */
 public interface HubPayloadFilter {
