+++ amplify-core/src/main/java/com/amplifyframework/storage/StorageAccessLevel.java
@@ -23,6 +23,6 @@ package com.amplifyframework.storage;
 public enum StorageAccessLevel {
     PUBLIC,
     PROTECTED,
-    PRIVATE;
+    PRIVATE
 }
