--- amplify-core/src/main/res/raw/amplifyconfiguration.json
@@ -1,61 +1,0 @@
-{
-  "UserAgent": "aws-amplify-cli/2.0",
-  "Version": "1.0",
-  "Storage": {
-    "plugins": {
-      "AWSS3StoragePlugin": {
-        "Bucket": "my-s3-bucket",
-        "Region": "us-west-2",
-        "TransferAcceleration": true
-      }
-    }
-  },
-  "Analytics": {
-    "plugins": {
-      "AmazonPinpointAnalyticsPlugin": {
-        "AppId": "xx113344",
-        "Region": "us-east-1",
-        "optOut": "ALL",
-        "globalUserAttributes": {
-          "interests": [
-            "football",
-            "basketball",
-            "AWS"
-          ]
-        }
-      },
-      "AmazonKinesisAnalyticsPlugin": {
-        "Stream": "my-kinesis-stream",
-        "PartitionKey": "my-kinesis-partition-key",
-        "Region": "us-east-1"
-      }
-    }
-  },
-  "API": {
-    "plugins": {
-      "AWSAPIPlugin": {
-        "friendly-name-1": {
-          "Endpoint": "http://api-gw-endpoint-1",
-          "Region": "us-east-1"
-        },
-        "friendly-name-2": {
-          "Endpoint": "http://api-gw-endpoint-2",
-          "Region": "us-east-2"
-        },
-        "friendly-name-3": {
-          "Endpoint": "http://graphql-endpoint-1",
-          "Region": "us-east-1",
-          "ApiKeyProvider": "DefaultAPIKeyProvider"
-        }
-      },
-      "FirebaseRESTfulAPI": {
-        "endpoint": "http://firebase-gw-endpoint-1.firebaseurl.com/v1/photos",
-        "secret": "ABC123",
-        "supportedOperations": [
-          "GET",
-          "POST"
-        ]
-      }
-    }
-  }
-}
