+++ amplify-core/src/main/res/values/strings.xml
@@ -14,6 +14,4 @@
    permissions and limitations under the License.
 -->
 
-<resources>
-    <string name="app_name">aws-amplify-core</string>
-</resources>
+<resources />
