+++ aws-amplify-analytics-pinpoint/src/main/res/values/strings.xml
@@ -14,6 +14,4 @@
    permissions and limitations under the License.
 -->
 
-<resources>
-    <string name="app_name">Amplify Analytics Plugin for Amazon Pinpoint</string>
-</resources>
+<resources />
